🗃Refactor: Component 리팩토링
