+++ src/components/index.ts
@@ -1,10 +1,10 @@
 export * from '@/components/core/Component';
 
-export * from '@/components/cafe/CafeHeader';
-export * from '@/components/cafe/CafeMenu';
-export * from '@/components/cafe/CafeModal';
+export * from '@/components/cafe/Header';
+export * from '@/components/cafe/Menu';
+export * from '@/components/cafe/Modal';
 
-export * from '@/components/cafe/CafeOrderList';
-export * from '@/components/cafe/CafeOrderListRow';
+export * from '@/components/cafe/OrderList';
+export * from '@/components/cafe/OrderListRow';
 
-export * from '@/components/cafe/CafeServing';
+export * from '@/components/cafe/Serving';
